// packages/nx-prisma/src/generators/configuration/schema.d.ts
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface ConfigurationGeneratorSchema {
  /**
   * The database provider for your schema.
   */
  database?: 'cockroachdb' | 'mongodb' | 'mysql' | 'postgresql' | 'sqlite' | 'sqlserver';
  /**
   * The name of the project to add the Prisma setup to.
   */
  project: string;
  /**
   * A directory where the prisma schema is placed.
   */
  directory?: string;
  /**
   * Skips formatting the workspace after the generator completes.
   */
  skipFormat?: boolean;
  /**
   * Do not add dependencies to `package.json`
   */
  skipPackageJson?: boolean;
}